requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
passlib==1.7.4
mysql-connector==2.2.9
mysql-connector-python==8.4.0
requests
streamlit-option-menu

